+++ tests/eeprom_test.cpp
@@ -4,12 +4,12 @@
 
 class TestingEepromDriver : public emb::eeprom::DriverInterface {
 private:
-    static const size_t _page_bytes = 64;
-    static const size_t _page_count = 8;
+    static const int _page_bytes = 64;
+    static const int _page_count = 8;
     static uint16_t _data[_page_count][_page_bytes/2];
 public:
-    virtual emb::eeprom::Error read(uint16_t page, uint16_t addr, uint8_t* buf, size_t len, emb::chrono::milliseconds timeout) {
-        for (size_t i = 0; i < len; ++i) {
+    virtual emb::eeprom::Error read(uint16_t page, uint16_t addr, uint8_t* buf, int len, emb::chrono::milliseconds timeout) {
+        for (int i = 0; i < len; ++i) {
             if (((addr + i) % 2) == 0) {
                 buf[i] = _data[page][(addr + i)/2] & 0x00FF;
             } else {
@@ -19,8 +19,8 @@ public:
         return emb::eeprom::Error::none;
     }
 
-    virtual emb::eeprom::Error write(uint16_t page, uint16_t addr, const uint8_t* buf, size_t len, emb::chrono::milliseconds timeout) {
-        for (size_t i = 0; i < len; ++i) {
+    virtual emb::eeprom::Error write(uint16_t page, uint16_t addr, const uint8_t* buf, int len, emb::chrono::milliseconds timeout) {
+        for (int i = 0; i < len; ++i) {
             if (((addr + i) % 2) == 0) {
                 _data[page][(addr + i)/2] = (_data[page][(addr + i)/2] & 0xFF00) | buf[i];
             } else {
@@ -30,8 +30,8 @@ public:
         return emb::eeprom::Error::none;
     }
 
-    virtual size_t page_bytes() const { return _page_bytes; }
-    virtual size_t page_count() const { return _page_count; }
+    virtual int page_bytes() const { return _page_bytes; }
+    virtual int page_count() const { return _page_count; }
 };
