+++ tests/filter_test.cpp
@@ -4,7 +4,7 @@
 void emb::tests::filter_test() {
     /* MovingAvgFilter */
     emb::movavg_filter<int, 5> mvavg_filter_i16;
-    for (size_t i = 1; i <= 5; ++i) {
+    for (int i = 1; i <= 5; ++i) {
         mvavg_filter_i16.update(i);
     }
     EMB_ASSERT_EQUAL(mvavg_filter_i16.output(), 3);
@@ -18,7 +18,7 @@ void emb::tests::filter_test() {
 
     emb::array<float, 10> filter_array;
     emb::movavg_filter<float, 10> mvavg_filter_f32(filter_array);
-    for (size_t i = 0; i < mvavg_filter_f32.size(); ++i) {
+    for (int i = 0; i < mvavg_filter_f32.size(); ++i) {
         mvavg_filter_f32.update(emb::numbers::pi * (1 + (i % 2)));
     }
     EMB_ASSERT_EQUAL(mvavg_filter_f32.output(), emb::numbers::pi * 1.5f);
