+++ tests/math_test.cpp
@@ -37,7 +37,7 @@ void emb::tests::math_test() {
 
     emb::integrator<float, float> integ(0.1f, emb::range<float>(-10.f, 10.f), -20.f);
     EMB_ASSERT_EQUAL(integ.output(), -10.f);
-    for (size_t i = 0; i < 50; ++i) {
+    for (int i = 0; i < 50; ++i) {
         integ.update(3);
     }
     EMB_ASSERT_TRUE(emb::range<float>(4.999f, 5.001f).contains(integ.output()));
@@ -45,7 +45,7 @@ void emb::tests::math_test() {
     EMB_ASSERT_TRUE(emb::range<float>(-4.999f, -5.001f).contains(integ.output()));
     integ.reset();
     EMB_ASSERT_EQUAL(integ.output(), -10.f);
-    for (size_t i = 0; i < 5000; ++i) {
+    for (int i = 0; i < 5000; ++i) {
         integ.update(1000);
     }
     EMB_ASSERT_EQUAL(integ.output(), 10.f);
