+++ tests/queue_test.cpp
@@ -20,7 +20,7 @@ void emb::tests::queue_test() {
     EMB_ASSERT_TRUE(queue.empty());
     EMB_ASSERT_TRUE(!queue.full());
 
-    for (size_t i = 1; i <= queue.capacity(); ++i) {
+    for (int i = 1; i <= queue.capacity(); ++i) {
         queue.push(i);
         EMB_ASSERT_EQUAL(queue.back(), i);
     }
@@ -38,23 +38,23 @@ void emb::tests::queue_test() {
 
 
     queue.clear();
-    for (size_t i = 1; i <= queue.capacity(); ++i) {
+    for (int i = 1; i <= queue.capacity(); ++i) {
         queue.push(i);
     }
-    for (size_t i = 0; i < queue.capacity()/2; ++i) {
+    for (int i = 0; i < queue.capacity()/2; ++i) {
         queue.pop();
     }
     EMB_ASSERT_EQUAL(queue.size(), 5);
     EMB_ASSERT_EQUAL(queue.front(), 6);
     EMB_ASSERT_EQUAL(queue.back(), queue.capacity());
-    for (size_t i = 0; i < queue.capacity()/2; ++i) {
+    for (int i = 0; i < queue.capacity()/2; ++i) {
         queue.pop();
     }
     EMB_ASSERT_EQUAL(queue.size(), 0);
     EMB_ASSERT_TRUE(queue.empty());
 
     queue.clear();
-    for (size_t i = 1; i <= queue.capacity() + 5; ++i) {
+    for (int i = 1; i <= queue.capacity() + 5; ++i) {
         if (!queue.full()) queue.push(i);
     }
     EMB_ASSERT_EQUAL(queue.size(), 10);
@@ -68,12 +68,12 @@ void emb::tests::queue_test() {
     EMB_ASSERT_EQUAL(queue.front(), 3);
 
     queue.clear();
-    for (size_t i = 1; i <= queue.capacity() + 5; ++i) {
+    for (int i = 1; i <= queue.capacity() + 5; ++i) {
         if (!queue.full()) {
             queue.push(i);
         }
     }
-    for (size_t i = 1; i <= queue.capacity(); ++i) {
+    for (int i = 1; i <= queue.capacity(); ++i) {
         EMB_ASSERT_EQUAL(queue.front(), i);
         queue.pop();
     }
